Temp linechart

+++ src/CMPUT301W21T23SmartDataBook/app/src/main/java/com/example/cmput301w21t23_smartdatabook/database/Database.java
@@ -241,7 +241,7 @@ public class Database {
 											}
 
 										} else {
-											temp = Float.parseFloat(document.get("Trial Value").toString());
+											temp = Double.parseDouble(document.get("Trial Value").toString());
 										}
 										add(temp);
 										add(document.get("StringDate").toString());
